refactor: replace useMarketDetail with useAutoRefreshTokenDetail in MarketDetailV2
- Removed the useMarketDetail hook and its initialization logic.
- Introduced useAutoRefreshTokenDetail for automatic token detail fetching every 10 seconds.
- Updated MarketDetailV2 to utilize the new hook, enhancing data refresh capabilities and simplifying the component structure.
