+++ packages/kit/src/views/Market/MarketDetailV2/hooks/useAutoRefreshTokenDetail.ts
@@ -1,0 +1,32 @@
+import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
+import { useTokenDetailActions } from '@onekeyhq/kit/src/states/jotai/contexts/marketV2';
+
+interface IUseMarketDetailDataProps {
+  tokenAddress: string;
+  networkId: string;
+}
+
+export function useAutoRefreshTokenDetail(data: IUseMarketDetailDataProps) {
+  const { current: tokenDetailActions } = useTokenDetailActions();
+
+  return usePromiseResult(
+    async () => {
+      // Set the tokenAddress and networkId in jotai state
+      tokenDetailActions.setTokenAddress(data.tokenAddress);
+      tokenDetailActions.setNetworkId(data.networkId);
+
+      // Fetch token detail data
+      await tokenDetailActions.fetchTokenDetail(
+        data.tokenAddress,
+        data.networkId,
+      );
+    },
+    [data.tokenAddress, data.networkId, tokenDetailActions],
+    {
+      watchLoading: true,
+      pollingInterval: 10_000,
+      revalidateOnFocus: true,
+      revalidateOnReconnect: true,
+    },
+  );
+}
